% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_bic_overview.R
\name{aic_bic_overview}
\alias{aic_bic_overview}
\title{Create an overview table for the AIC and BIC}
\usage{
aic_bic_overview(
  mod_list,
  data_list,
  cov_types = rep("iid", length(mod_list)),
  return_ll = TRUE
)
}
\arguments{
\item{mod_list}{list of fitted models}

\item{data_list}{data list used to generate the models}

\item{cov_types}{Used covariance assumtion.
"iid" (No correlation and identical variance),
"classic" (pearson covariance estimation),
"ls" (shrinkage covariance estimation),
"var" no correlation but different variances.
@param return_ll Should the log likelihood also be returned?}
}
\value{
data.frame
}
\description{
Create an overview table for the AIC and BIC
}
