% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backshift_matrix.R
\name{backshift_matrix}
\alias{backshift_matrix}
\title{Quantile-Comparison Plot between data and a simulation-list.}
\usage{
backshift_matrix(coeff, n = 21L)
}
\arguments{
\item{coeff}{coefficients}

\item{n}{matrix dimension}
}
\value{
Backshift-Matrix
}
\description{
Quantile-Comparison Plot between data and a simulation-list.
}
\details{
the first n-1 values in coeff represent the values of the subdiagional the next n-2 for the sub-subdiagnol and so on
}
\examples{
backshift_matrix(rnorm(5+4+3), 6L)

}
