% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_nlr.R
\name{bayes_nlr}
\alias{bayes_nlr}
\title{Sample from a bayesian Nonlinear Regression}
\usage{
bayes_nlr(
  formula,
  data,
  start,
  log_poster = log_poster_gauss,
  hyperpara,
  cov = T,
  suggest_sample = function(x) {
     rnorm(length(x), x, sd = 1)
 },
  n_iter = 1000,
  verbose = T
)
}
\arguments{
\item{formula}{a model formula including variables and parameters.}

\item{data}{a data frame containing data}

\item{start}{a named list of starting values}

\item{log_poster}{a function which is propotional to the logarithmic posteriori density}

\item{hyperpara}{a named list of hyperparameters. See details!}

\item{cov}{if TRUE the covariance is estimated using the inverse Wishard distribution
if FALSE only the variances are estimated via inverse gamma.SEE details!}

\item{suggest_sample}{a function to sample from a suggested distribution}

\item{n_iter}{number of iterations}
}
\value{
posteriori samples
}
\description{
Sample from a bayesian Nonlinear Regression
}
\details{
if cov == TRUE "hyperpara" must contain the covariance matrix \eqn{V} and \eqn{k} for \eqn{IW(kV,k)} prior.
if cov == FALSE "hyperpara" must contain shape and rate
}
\examples{
exa_func <- function(t, a,b,c,d){
  a - b/2 + pmin(abs(a)*t, c- abs(b)*t, d)
}

test_df <- data.frame(t = rep(1:20, 24))

test_df$y <- exa_func(test_df$t, 3,2,40,15) + rnorm(20*24, sd = 1)


bayes_nlr(y~exa_func(t, a,b,c,d),
   data = test_df,
   start = c("a"= 2,"b"=2, c=40, d = 15),
  log_poster = log_poster_gauss,
  hyperpara = list("a"=0,"b"=10, "V" = diag(20), "k" = 1),
  n_iter = 10
  )

}
