% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estimation.R
\name{cov_estimation}
\alias{cov_estimation}
\title{Different types of covariance estimation}
\usage{
cov_estimation(
  data,
  cov_type = c("sd", "sample", "SSS", "LW"),
  penalty = 1,
  alpha = 1/100,
  ...
)
}
\arguments{
\item{data}{matrix or data.frame of dimension I times J}

\item{cov_type}{covariance matrix type.
sd: only include variance / standard deviation.
samlple: estimate sample coviarance matrix.
SSS: Simple sample shrinkage: \eqn{\Sigma}(1-\eqn{\alpha}) + D\eqn{\alpha}, D=diag(\eqn{\sigma_1^2},\dots, \eqn{\sigma_J^2})
LW: Ledoit-Wolf linear shrinkage Estimation. See cvCovEst::linearShrinkLWEst}

\item{penalty}{nrow(data)-penalty degrees of freedom}

\item{alpha}{in Simple Sample Shrink (SSS). Otherwise ignored.}

\item{...}{further arguments for cov or var}
}
\description{
Different types of covariance estimation
}
