% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_list_to_intervall.R
\name{cov_list_to_intervall}
\alias{cov_list_to_intervall}
\title{Converts a list of covariance matrices into a matrix representing component-wise interval estimates}
\usage{
cov_list_to_intervall(x, alpha = 0.05, df_names = NULL)
}
\arguments{
\item{x}{a list of covariance matrices}

\item{alpha}{significance niveau}

\item{df_names}{row names for data.frame}
}
\description{
Converts a list of covariance matrices into a matrix representing component-wise interval estimates
}
