% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_list_to_var_df.R
\name{cov_list_to_var_df}
\alias{cov_list_to_var_df}
\title{Transforms a list of covariances into a data.frame of variances}
\usage{
cov_list_to_var_df(x, df_names = NULL)
}
\arguments{
\item{x}{list of matrices}

\item{df_names}{row names for data.frame}
}
\description{
Transforms a list of covariances into a data.frame of variances
}
\examples{
a <- replicate(
   1000, expr = {
   x <- rnorm(10)
   x <- x\%*\%t(x)
   x
   },
   simplify = F
   )

cov_list_to_var_df(a)
}
