% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrunc2.R
\name{dtrunc2}
\alias{dtrunc2}
\title{Like dtrunc from the LaplacesDemon Package but it can deal with values outside of the range...}
\usage{
dtrunc2(x, spec, a = -Inf, b = Inf, log = FALSE, ...)
}
\description{
Like dtrunc from the LaplacesDemon Package but it can deal with values outside of the range...
}
