% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_multi_gauss.R
\name{entropy_multi_gauss}
\alias{entropy_multi_gauss}
\alias{entropy_multi_gauss.nls}
\alias{entropy_multi_gauss.matrix}
\alias{entropy_multi_gauss.list}
\title{entropy_multi_gauss:}
\usage{
entropy_multi_gauss(x)

\method{entropy_multi_gauss}{nls}(x)

\method{entropy_multi_gauss}{matrix}(x)

\method{entropy_multi_gauss}{list}(x)
}
\arguments{
\item{x}{A nls model or a matrix or a list}
}
\description{
\eqn{J(\log(2\pi)+1)/2 + \log|\Sigma|/2}
}
\details{
A details of entropy_multi_gauss
}
