% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_range.R
\name{expected_range}
\alias{expected_range}
\title{Expected range of \eqn{J} random variables \eqn{log(X)}, where \eqn{X} follows a \eqn{Ga(N/2,1)} distribution}
\usage{
expected_range(N, J)
}
\arguments{
\item{N}{degrees of freedom}

\item{J}{Number of observations}
}
\description{
Expected range of \eqn{J} random variables \eqn{log(X)}, where \eqn{X} follows a \eqn{Ga(N/2,1)} distribution
}
\details{
Same functions exits for min and max. Called expected_min and expected_max.
}
