% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_model_information.R
\name{final_model_information}
\alias{final_model_information}
\title{Informations about the final Version(s) of the sound reduction index model!}
\usage{
final_model_information()
}
\description{
Informations about the final Version(s) of the sound reduction index model!
}
