% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_info_tab.R
\name{fit_info_tab}
\alias{fit_info_tab}
\alias{fit_info_tab.nls}
\alias{fit_info_tab.list}
\title{fit_info_tab: Title}
\usage{
fit_info_tab(mod, ...)

\method{fit_info_tab}{nls}(mod, digits = NULL, freq_subset = FALSE)

\method{fit_info_tab}{list}(x, ...)
}
\arguments{
\item{...}{other arguments like digits}

\item{x}{A model or a list}
}
\value{
\code{NULL}

\code{NULL}
}
\description{
A description of fit_info_tab
}
\details{
A details of fit_info_tab
}
