% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_basic.R
\name{gg_basic}
\alias{gg_basic}
\title{Easy frequency depending plots}
\usage{
gg_basic(
  data,
  mapping = aes(),
  ymin = 10,
  ymax = 90,
  sound = T,
  lab = T,
  isl = F,
  english = T
)
}
\arguments{
\item{data}{data.frame}

\item{mapping}{similar to ggplot}

\item{ymin}{lower ylim value}

\item{ymax}{upper ylim value}

\item{sound}{Is the curve a sound reduction index R or an impact sound level curve?}

\item{lab}{Is it labatory data? (Ignored if sound == FALSE)}

\item{isl}{Is it impact sound level? (Ignored if sound == FALSE)}

\item{english}{Should the axis labeling be in English?}
}
\description{
Easy frequency depending plots
}
