% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggWhiteRes.R
\name{ggWhiteRes}
\alias{ggWhiteRes}
\title{Whitened residual plots}
\usage{
ggWhiteRes(
  df,
  mapping = aes(),
  numbCurves = nrow(df)/numbVariants/21,
  numbVariants = 1L,
  txtsize = 2.8,
  rmargin = 80,
  tikz = F
)
}
\arguments{
\item{df}{data.frame with whintened residuals}

\item{mapping}{see ?ggplot2::ggplot}

\item{numbCurves}{number of curves}

\item{numbVariants}{number of variants}

\item{txtsize}{text size of line informations}

\item{rmargin}{size of the right margin}

\item{tikz}{If TRUE, the percent signs are replaced with p because it is more convenient for TikZ.}
}
\description{
Whitened residual plots
}
