% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_overview.R
\name{ic_overview}
\alias{ic_overview}
\title{Create an overview table for the AIC and BIC}
\usage{
ic_overview(modlist)
}
\arguments{
\item{modlist}{list of fitted models}
}
\value{
data.frame
}
\description{
Create an overview table for the AIC and BIC
}
