% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_sound_function.R
\name{impact_sound_raw}
\alias{impact_sound_raw}
\title{Calculate the impact sound on a raw ceiling acording to DIN EN ISO 12354-2}
\usage{
impact_sound_raw(
  thickness,
  density,
  volume,
  refrence_freq = 1000,
  offset = 155,
  cL = 3500,
  c0 = 340,
  freq = NULL,
  max_rl = 1.18,
  max_rl_type = "value"
)
}
\arguments{
\item{volume}{room volumne}

\item{offset}{offset (= 155 dB in standard)}

\item{cL}{Longitudinal velocity}

\item{c0}{Sound speed in air in m/s. Default = 340, which is the speed at 15 °C}

\item{freq}{Frequency in Hz}

\item{max_rl}{See ?smooth_radiation_level}

\item{max_rl_type}{See ?smooth_radiation_level}

\item{mass}{mass per unit area}

\item{reference_freq}{reference frequency}
}
\description{
Equation (B.2) from DIN EN ISO 12354-2
}
