% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_sound_function.R
\name{impact_sound_reduction}
\alias{impact_sound_reduction}
\title{Calculate the impact sound reduction according to DIN EN ISO 12354-2}
\usage{
impact_sound_reduction(prefactor = 30, resonance_freq, freq = NULL)
}
\arguments{
\item{prefactor}{30 in (C.1) and 40 in (C.3)}

\item{freq}{Frequency in Hz}

\item{reference_freq}{reference frequency}
}
\description{
Equation (C.1) and (C.2) from DIN EN ISO 12354-2
}
