% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_sound_function.R
\name{impact_sound_standard}
\alias{impact_sound_standard}
\title{Calculate the impact sound according to DIN EN ISO 12354-2}
\usage{
impact_sound_standard(
  thickness,
  reduction_factor,
  resonance_freq,
  volume,
  density = 2300,
  offset = 155,
  freq = NULL,
  ...
)
}
\arguments{
\item{thickness}{wall thickness in m}

\item{reduction_factor}{strength of reduction: 30 in (C.1) and 40 in (C.3)}

\item{resonance_freq}{resonance frequency of the system in Hz}

\item{volume}{volume of the receiving room, in cubic metres.}

\item{density}{density of the floor, in kilograms per cubic metre;}

\item{offset}{additive dB shift. Equal to 155 dB in equation (B.2)}

\item{freq}{Frequency in Hz}

\item{...}{Further arguments for "impact_sound_raw"}
}
\description{
Difference between equation (B.2) and (C.1) from DIN EN ISO 12354-2
}
