% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_squares_lag_mat.R
\name{least_squares_lag_mat}
\alias{least_squares_lag_mat}
\title{Transforms a \eqn{I \times J} matrix into a LSL-Matrix}
\usage{
least_squares_lag_mat(y, maxlag)
}
\arguments{
\item{y}{\eqn{I \times J} matrix}

\item{maxlag}{maximum number of lags}
}
\description{
result is a \eqn{I (J-1)\times Q(maxlag)} matrix where

\eqn{Q(l) = l(J-(l+1)/2)}
}
