% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_continuation.R
\name{linear_continuation}
\alias{linear_continuation}
\title{Linear continuation as soon as a limit is undershot.}
\usage{
linear_continuation(y, limit = 10, orientation = 5)
}
\arguments{
\item{y}{Data vector}

\item{limit}{Limit value for y at which the values are no longer credible}

\item{orientation}{Number of observations used to continue the data linearly.}
}
\description{
Linear continuation as soon as a limit is undershot.
}
