% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_poster_gauss.R
\name{log_poster_gauss}
\alias{log_poster_gauss}
\title{Function which is proportional to logarithmic gaussian distribution}
\usage{
log_poster_gauss(
  var_values,
  target_values,
  Sigma,
  para_names,
  hyperpara = c(a = 0, b = 1),
  formu
)
}
\arguments{
\item{var_values}{named list of variable values}

\item{target_values}{vector of response variable}

\item{Sigma}{Covariance matrix}

\item{para_names}{vector of variable names}

\item{hyperpara}{vector or list of hyperparameters}

\item{formu}{formula}
}
\description{
Function which is proportional to logarithmic gaussian distribution
}
