% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike_mcmc_spm.R
\name{loglike_mcmc_spm}
\alias{loglike_mcmc_spm}
\title{Calculate A S by N log Likelihood matrix}
\usage{
loglike_mcmc_spm(fit_curve_list, cov_list, target, cov = T)
}
\arguments{
\item{fit_curve_list}{a list of mcmc fit curves of length S}

\item{cov_list}{a list of mcmc covariances of length S}

\item{target}{a target vector of length N}

\item{cov}{TRUE if correlation should be included}
}
\description{
Calculate A S by N log Likelihood matrix
}
\details{
Result can be used for loo:loo and loo::waic
}
