% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_cv_data.R
\name{loo_cv_data}
\alias{loo_cv_data}
\title{leave one out cross validation on the data under gaussian assumption}
\usage{
loo_cv_data(
  data,
  cov_type = c("sd", "sample", "SSS", "LW"),
  penalty = 1,
  alpha = 1/100
)
}
\arguments{
\item{data}{matrix or data.frame of dimension I times J}

\item{cov_type}{type of covariance estimation: See ?cov_estimation}

\item{penalty}{See ?cov_estimation}

\item{alpha}{Only for SSS. See ?cov_estimation}
}
\description{
leave one out cross validation on the data under gaussian assumption
}
\details{
Calculates the negative log density for each element of the group
where the parameters are estimated with the remaining observations! Lower values are better!

The expected value is equal to the entropy!


-log(N(\eqn{y_i - \bar{y_{-i}}, \hat{\Sigma}_{-i}}))
}
