% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_cv_nlsLMcov.R
\name{loo_cv_nlsLMcov}
\alias{loo_cv_nlsLMcov}
\title{leave one out cross validation for nlsLMcov function}
\usage{
loo_cv_nlsLMcov(
  formula,
  data,
  start,
  groups,
  cov_type = c("sd", "sample", "SSS", "LW", "iid", "tvAR"),
  verbose = TRUE,
  maxlag = 1,
  curve_len = 21L,
  ...
)
}
\arguments{
\item{formula}{nonlinear model formula like nlsLM}

\item{data}{data.frame}

\item{start}{a named list or named numeric vector of starting values.}

\item{groups}{column name which indicates the groups}

\item{cov_type}{type of covariance estimation: See ?nlsLMcov}

\item{verbose}{If TRUE, it is communicated for which curves the calculations have already been completed.}

\item{maxlag}{maximum number of lags if cov_type is set to 'tvAR'}

\item{curve_len}{length of measurement curves}

\item{...}{further arguments of nlsLMcov}
}
\description{
leave one out cross validation for nlsLMcov function
}
\details{
Calculates the negative log density for each element of the group
where the parameters are estimated with the remaining observations! Lower values are better!

The expected value is equal to the entropy!


-log(N(\eqn{y_i - g(\hat{\theta}_{-i}), \hat{\Sigma}_{-i}}))
}
