% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis_dist.R
\name{mahalanobis_dist}
\alias{mahalanobis_dist}
\title{Mahalanobis Distance}
\usage{
mahalanobis_dist(x, center, cov, inverted = FALSE, chol = FALSE, ...)
}
\arguments{
\item{x}{p dimensional vector}

\item{center}{p dimensional mean vector}

\item{cov}{\eqn{p\times p} covariance matrix}

\item{inverted}{if TRUE, cov is supposed to contain the inverse of the covariance matrix.}

\item{chol}{if TRUE cov is a lower triangle matrix.}

\item{...}{further arguments for solve if inverted = FALSE}
}
\value{
Mahalanobis Distance
}
\description{
Mahalanobis Distance
}
