% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis_samples.R
\name{mahalanobis_samples}
\alias{mahalanobis_samples}
\title{Mahalanobis Distance for MCMC, Bootstrap or other samples}
\usage{
mahalanobis_samples(
  y_samples,
  cov_samples,
  cholesky = FALSE,
  inverse = FALSE,
  centered = FALSE
)
}
\arguments{
\item{y_samples}{\eqn{S\times p} samples}

\item{cov_samples}{\eqn{S\times p\times p} samples of covariance matrices}

\item{cholesky}{TRUE if matrices are lower triangle matrices}

\item{centered}{TRUE if y_samples are already centered otherwise the samples will be centered}

\item{inverted}{TRUE if matrices are inverse matrices}
}
\value{
Mahalanobis distance for samples
}
\description{
Mahalanobis Distance for MCMC, Bootstrap or other samples
}
