% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCqqPlots.R
\name{MCMCqqPlotsChi}
\alias{MCMCqqPlotsChi}
\title{Converts a \eqn{S \times N} matrix containing chi-squared-distributed samples into a dataframe for qq plots.}
\usage{
MCMCqqPlotsChi(x, df, numb_q = nrow(x), index = NULL)
}
\arguments{
\item{x}{\eqn{S \times N} matrix}

\item{df}{degree of freedom for Chisq}

\item{numb_q}{number of quantiles}

\item{index}{column names of x}
}
\description{
Converts a \eqn{S \times N} matrix containing chi-squared-distributed samples into a dataframe for qq plots.
}
