% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsLMcov.R
\name{nlsLMcov}
\alias{nlsLMcov}
\title{minpack.lm::nlsLM including a (co)variance estimation}
\usage{
nlsLMcov(
  formula,
  data,
  start,
  cov_type = c("sd", "sample", "SSS", "LW", "iid", "tvAR"),
  n_iter = 3,
  sss_alpha = 1/10000,
  maxlag = 1,
  data_cov = F,
  curve_len = 21L,
  ...
)
}
\arguments{
\item{formula}{nonlinear model formula like nlsLM}

\item{data}{data.frame or list.}

\item{start}{a named list or named numeric vector of starting values.}

\item{cov_type}{\itemize{
 \item sd: only include variance / standard deviation.
 \item samlple: estimate sample coviarance matrix.
 \item SSS: \eqn{\Sigma}(1-\eqn{\alpha}) + D\eqn{\alpha}, D=diag(\eqn{\sigma_1^2},\dots, \eqn{\sigma_J^2})
 \item LW: Ledoit-Wolf linear shrinkage Estimation. See cvCovEst::linearShrinkLWEst
 \item iid: iid error which is equal to nlsLM
 \item tvAR: time varying autoregressive covariance structure
}}

\item{n_iter}{number of iterations for covariance estimation. (Gets ignored if data_cov=TRUE)}

\item{sss_alpha}{\eqn{\alpha} in Simple Sample Shrink (SSS). Otherwise ignored.}

\item{maxlag}{maximum number of lags for tvAR assumption}

\item{data_cov}{If TRUE the covariance matrix is centered by using the mean values}

\item{curve_len}{length of measurement curves}

\item{...}{further arguments for nlsLM}
}
\description{
minpack.lm::nlsLM including a (co)variance estimation
}
\details{
All covariances use: I-k degrees of freedom.
Where I is the number of measurement curves an k the number of fit parameters.

First, a model without covariance assumption is estimated.
Then the residuals of this model are used for the covariance estimation.
This process is repeated n_times.
}
