% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsLMtvAR.R
\name{nlsLMtvAR}
\alias{nlsLMtvAR}
\title{minpack.lm::nlsLM including tvAR coefficients}
\usage{
nlsLMtvAR(formula, data, start, max_lag = 1L, n_iter = 3L, data_cov = F, ...)
}
\arguments{
\item{formula}{nonlinear model formula like nlsLM}

\item{data}{data.frame or list.}

\item{start}{a named list or named numeric vector of starting values.}

\item{max_lag}{maximum number of lags}

\item{n_iter}{number of iterations for covariance estimation. (Gets ignored if data_cov=TRUE)}

\item{data_cov}{If TRUE the covariance matrix is centered by using the mean values}

\item{...}{further arguments for nlsLM}
}
\description{
minpack.lm::nlsLM including tvAR coefficients
}
