% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntp_data.R
\docType{data}
\name{ntp_data}
\alias{ntp_data}
\title{Impact sound insulation}
\format{
\describe{
   \item{NTP_j.Hz}{Impact sound level measurement at \eqn{j} Hz.}
   \item{seltsam_mbbm}{TRUE if the curve is strange, according to ‘Andreas Meier’ from Müller BBM.}
 }
data(ntp_data)
}
\usage{
ntp_data
}
\description{
Note: These are NOT actual measurements but artificially generated sample data.
For actual measurements, use the \emph{SoundModelingData} package if available!
}
\keyword{datasets}
