% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutOfRange.R
\name{OutOfRange}
\alias{OutOfRange}
\title{Checks if values are out of range}
\usage{
OutOfRange(
  df,
  targetName,
  lowerName,
  upperName,
  result = c("data.frame", "sum", "mean", "matrix"),
  ncol = 21L
)
}
\arguments{
\item{df}{a data.frame with lower and upper bounds}

\item{targetName}{name of the target column}

\item{lowerName}{name of the lower column}

\item{upperName}{name of the upper column}

\item{result}{result format see details}

\item{ncol}{column length for the case: result == matrix}
}
\description{
Checks if values are out of range
}
\details{
\itemize{
 \item data.frame: Return is the original data.frame including a boolean column.
  TRUE means the target value is not in the range
 \item sum: Returns the absolute number of values outside of the range
 \item mean: Returns the relative number of values outside of the range
 \item matrix: Returns a boolean matrix
}
}
