% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_range.R
\name{prediction_range}
\alias{prediction_range}
\title{Pointwise prediction interval}
\usage{
prediction_range(
  samples,
  alpha = 0.05,
  qnames = c("lower", "upper"),
  simultaneous = FALSE,
  data = NULL
)
}
\arguments{
\item{samples}{a \eqn{S \times I\times J} array of simulations}

\item{alpha}{The range between \eqn{\alpha/2} and \eqn{1-\alpha/2} is calculated}

\item{qnames}{Column names for the quantiles}

\item{simultaneous}{if TRUE alpha is adjusted for simultanious intervals}

\item{data}{If NULL, only the range is returned, otherwise cbind is also performed}
}
\description{
Uses simulations/samples to calculate a point-by-point prediction range
}
