% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot_sim.R
\name{qq_plot_sim}
\alias{qq_plot_sim}
\title{Quantile-Comparison Plot between data and a simulation-list.}
\usage{
qq_plot_sim(x, simulation)
}
\arguments{
\item{x}{a data vector}

\item{simulation}{a matrix of simulations}
}
\value{
qq-plot
}
\description{
Quantile-Comparison Plot between data and a simulation-list.
}
\examples{
qq_plot_sim(rnorm(24), t(replicate(1000, rnorm(24))))

}
