% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot_simulation.R
\name{qq_plot_simulation}
\alias{qq_plot_simulation}
\title{Quantile-Comparison Plot between data and a simulation-list.}
\usage{
qq_plot_simulation(data, simulation)
}
\arguments{
\item{data}{a data vector.}

\item{simulation}{a list of simulations.}
}
\value{
qq-plot
}
\description{
Quantile-Comparison Plot between data and a simulation-list.
}
\details{
THIS FUNCTION IS OUTDATED! Please use 'qq_plot_sim' instead!
}
\examples{
qq_plot_simulation(rnorm(24), replicate(1000, rnorm(24), simplify = F))

}
