% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_sim_value.R
\name{qq_sim_value}
\alias{qq_sim_value}
\title{Quantile-Comparison values between data and a simulation-list.}
\usage{
qq_sim_value(x, simulation)
}
\arguments{
\item{x}{a data vector}

\item{simulation}{a matrix of simulations}
}
\value{
\eqn{1-2|p-1/2|}, where  \eqn{p} is the estimated probablity that the observed quantile is smaller than the simulated quantile.
}
\description{
Quantile-Comparison values between data and a simulation-list.
}
\examples{
qq_sim_value(rnorm(24), t(replicate(1000, rnorm(24))))
}
