% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsLMcov.R
\name{resid_to_cov}
\alias{resid_to_cov}
\title{Transforms a matrix of residuals into a covariance matrix}
\usage{
resid_to_cov(resmat, covtype, k = 0, alpha = 1/100, maxlag = 1)
}
\arguments{
\item{resmat}{a matrix of residuals or other centered values}

\item{covtype}{covariance matrix type.
sd: only include variance / standard deviation.
samlple: estimate sample coviarance matrix.
SSS: Simple sample shrinkage: \eqn{\Sigma}(1-\eqn{\alpha}) + D\eqn{\alpha}, D=diag(\eqn{\sigma_1^2},\dots, \eqn{\sigma_J^2})
LW: Ledoit-Wolf linear shrinkage Estimation. See cvCovEst::linearShrinkLWEst
tvAR: time-varying autoregressive assumption with 'maxlag'}

\item{k}{nrow(resmat)-k degrees of freedom}

\item{maxlag}{maximum number of lags for tvAR assumption}
}
\description{
Transforms a matrix of residuals into a covariance matrix
}
