% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsLMcov.R
\name{residual_LWcov}
\alias{residual_LWcov}
\title{Transforms a matrix of residuals into a Ledoit-Wolf covariance matrix}
\usage{
residual_LWcov(resmat, k = 1)
}
\arguments{
\item{resmat}{a matrix of residuals or other centered values}

\item{k}{nrow(resmat)-k degrees of freedom}
}
\description{
Transforms a matrix of residuals into a Ledoit-Wolf covariance matrix
}
\details{
See ?cvCovEst::linearShrinkLWEst
}
