% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsLMcov.R
\name{residual_tvAR}
\alias{residual_tvAR}
\title{Transforms a matrix of residuals into a tvAR covariance matrix}
\usage{
residual_tvAR(resmat, maxlag = 1, k = 1, returnWhite = F)
}
\arguments{
\item{resmat}{a matrix of residuals or other centered values}

\item{maxlag}{maximum number of lags}

\item{k}{nrow(resmat)-k degrees of freedom}

\item{returnWhite}{return whitening matrix}

\item{curve_len}{length of measurement curves}
}
\description{
Transforms a matrix of residuals into a tvAR covariance matrix
}
