% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmoving_uniform.R
\name{rmoving_uniform}
\alias{rmoving_uniform}
\title{Sample from a moving uniform distribution}
\usage{
rmoving_uniform(x, a = -5, b = 5, l = sqrt(12))
}
\arguments{
\item{x}{Sample point}

\item{a}{minimum value}

\item{b}{maximum value}

\item{l}{length of searching area}
}
\description{
Sample from a moving uniform distribution
}
