% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_probabilities.R
\name{rw_probabilities}
\alias{rw_probabilities}
\title{Transforms a rw matrix into a dataframe which indicates the relative frequencies for minimum values.}
\usage{
rw_probabilities(rw, bounds = c(47, 50, 53, 56, 59), long = FALSE)
}
\arguments{
\item{rw}{\eqn{S \times T} matrix of Rw values. With \eqn{S} samples and number of wall thicknesses \eqn{T}.}

\item{bounds}{boundaries for Rw}

\item{long}{if TRUE the result data.frame is in long format.}
}
\description{
Calculates the Rw value
}
