% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_value.R
\name{rw_value}
\alias{rw_value}
\title{Calculate single value Rw}
\usage{
rw_value(r, digits = 1, min_s = -100, lvl = 1)
}
\arguments{
\item{r}{measured values}

\item{digits}{number of digits}

\item{min_s}{minimal shift}

\item{lvl}{relevant for recursive execution}
}
\value{
Optimal shift
}
\description{
Calculate single value Rw
}
\examples{
rw_value(rnorm(16))
}
