% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_reduce.R
\name{sample_reduce}
\alias{sample_reduce}
\alias{sample_reduce.default}
\alias{sample_reduce.data.frame}
\title{Reduce (Metropolis) samples}
\usage{
sample_reduce(sample, ...)

\method{sample_reduce}{default}(sample, accepted, len, type = c("burn", "reverse"))

\method{sample_reduce}{data.frame}(sample, accepted, len, type = c("burn", "reverse"))
}
\arguments{
\item{sample}{a list or data.frame of samples}

\item{accepted}{a logical vector}

\item{len}{length finaly accepted samples}

\item{type}{if "burn" then the first len will be removed and after that the accepted values will be used
and for type "reverse" the accepted ones are selected first and then the last of these are selected.}
}
\description{
Reduce (Metropolis) samples
}
