% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_value.R
\name{single_value}
\alias{single_value}
\title{Single value for a given shift}
\usage{
single_value(shift, r)
}
\arguments{
\item{shift}{todo}

\item{r}{measured values}
}
\value{
Single value
}
\description{
Single value for a given shift
}
\examples{
single_value(0, rnorm(16))
}
