% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldf.R
\docType{data}
\name{sldf}
\alias{sldf}
\title{Soundproofing measure data with calcium silicate construction.}
\format{
\describe{
   \item{Building}{Wall number}
   \item{RpdB}{Measured R in dB}
   \item{fpHz}{Frequency in Hz}
   \item{s}{Surface in m²}
   \item{u}{Perimeter in m}
 }
data(sldf)
}
\usage{
sldf
}
\description{
Note: These are NOT actual measurements but artificially generated sample data.
For actual measurements, use the \emph{SoundModelingData} package if available!
}
\keyword{datasets}
