% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_radiation_level.R
\name{smooth_radiation_level}
\alias{smooth_radiation_level}
\title{Generate the radiation level for the typical terz-values and smooth it via moving average and moving geometric-average}
\usage{
smooth_radiation_level(
  lower_rl_bound = 1,
  upper_rl_bound = 1,
  delta_bound = 0.5,
  max_rl = 2,
  max_rl_type = c("dynamic", "value", "fc", "sigma3"),
  c0 = 340,
  w_perimeter = 12.57,
  w_surface = 9.8,
  f_c = (4 * sqrt(3) * 340^2)/(pi * 1250),
  freq = NULL
)
}
\arguments{
\item{lower_rl_bound}{f_c * lower_rl_bound defines the first radiation level region}

\item{upper_rl_bound}{f_c * upper_rl_bound defines the last radiation level region}

\item{delta_bound}{Defines the region for which delta2 is not zero.}

\item{max_rl}{Maximum value reached by the radiation level. (Ignored for types: "fc" and "sigma3")}

\item{max_rl_type}{Maximum radiation level type.}

\item{c0}{Speed of sound}

\item{w_perimeter}{Perimeter of the wall}

\item{w_surface}{Surface of a wall}

\item{f_c}{Coincidence cutoff frequency}

\item{freq}{Frequencies - if NULL the typical 50, 63, 100, ..., 4000, 5000 values are used.}
}
\value{
smooth_radiation_level an data.frame
}
\description{
Generate the radiation level for the typical terz-values and smooth it via moving average and moving geometric-average
}
\details{
For max_rl_type = 'dynamic' the maximal sigma is equal to \eqn{\min(\sigma_3, \mathrm{max\_rl}, \sqrt{1+u f_c/(5c_0)})}

For max_rl_type = 'value' the maximal sigma is equal to \eqn{\mathrm{max\_rl}}

For max_rl_type = 'fc' the maximal sigma is equal to \eqn{f_c}

For max_rl_type = 'sigma3' the maximal sigma is equal to \eqn{\sigma_3}
}
