% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundproofing_measure.R
\name{soundproofing_measure}
\alias{soundproofing_measure}
\title{Calculate the soundproofing measure R}
\usage{
soundproofing_measure(
  internal_loss = 0.015,
  lab_loss_factor = 0.9,
  cL = 1250,
  lower_t_bound = 0.89,
  upper_t_bound = 1.4,
  type = c("Original", "Average", "Geometric", "Rindel", "A_Rindel"),
  c0 = 340,
  rho0 = 1.23,
  mass = 440,
  w_perimeter = 12.57,
  w_surface = 9.8,
  smoothing_power = 2/3,
  smoothing_direction = c("left", "both"),
  smoothing_log = TRUE,
  plateau_style = c("max", "sum", "mixed", "smooth", "tau0add", "tau0Lim", "tau0Complex",
    "tau0real", "test", "approx", "none"),
  wall_thickness = 0.25,
  rho = 1760,
  poisson = 0.25,
  cL1 = cL,
  cLplateau = c("1", "eff"),
  volume_rec_room = 50,
  surface_rec_room = 85,
  fix_between_rad = NULL,
  fce_gamma = NULL,
  break_point = F,
  pre_bp_slope = 3,
  add_flank_transmission = F,
  flank_impact = 1,
  flank_power = 1,
  flank_start = NULL,
  freq = NULL,
  gm_weight = 4,
  sts = FALSE,
  max_rl = 2,
  max_rl_type = c("dynamic", "value", "fc", "sigma3"),
  offset = 0,
  ...
)
}
\arguments{
\item{internal_loss}{Internal loss factor}

\item{lab_loss_factor}{Constant "C" respresenting the boundary loss C/sqrt(f)}

\item{cL}{Effective longitudinal velocity}

\item{lower_t_bound}{If f < lower_t_bound*fc, the first section applies for the transmittance}

\item{upper_t_bound}{If f > upper_t_bound*fc, the last section applies for the transmittance}

\item{type}{Choose whether the radiation level should be smoothed with moving average, geometric moving average or not at all.}

\item{c0}{Sound speed in air in m/s. Default = 340, which is the speed at 15 °C}

\item{rho0}{Density of air in kg/m³. Default = 1.23, which is the density at 15 °C}

\item{mass}{Mass per unit area in kg/m². Default = 440.}

\item{w_perimeter}{Wall perimeter.}

\item{w_surface}{Wall surface.}

\item{smoothing_power}{Smoothing strength. If zero, no smoothing is applied and if one, the two values are set identically by one uncertainty point.}

\item{smoothing_direction}{If "left" only the values around lower_t_bound fc are getting smothed and if "both" the values around upper_t_bound fc will also be smoothed.}

\item{smoothing_log}{If FALSE the distance between two frequencies is calculated via absolute distance and if TRUE the logarithm of the frequencies is used instead.}

\item{plateau_style}{What should the curve look like for extremely high frequencies? *Add description here as soon as I get to grips with this description and know how to make it clearer, etc.*}

\item{wall_thickness}{Constuction thickness of the wall in m.}

\item{rho}{Density of the material in kg/m³. Default = 1760.}

\item{poisson}{Poisson number / material value. Between 0 and 0.5.}

\item{cL1}{(Quasi) longitudinal velocity / material longitudinal velocity}

\item{cLplateau}{which longitudinal velocity should be used for tau-plateau? cL1 or cLeff?}

\item{volume_rec_room}{Reception room volume in m³.}

\item{surface_rec_room}{Total surface in reception room in m².}

\item{fix_between_rad}{Should the values between lower_t_bound fc and upper_t_bound fc be set to an fixed value? If no set fix_between_rad to NULL.}

\item{fce_gamma}{gamma for fceff. If gamma == NULL the DIN EN ISO 12354-1 calculation is used. Otherwiese gamma should be between 0.5 and 1.}

\item{break_point}{Insert a straight line before a break point. If FALSE no break point is use. For numeric the break point is set to this frequency and for NULL the breakpoint is at f11.}

\item{pre_bp_slope}{Slope of the line before break point. (Pre break point slope.)}

\item{add_flank_transmission}{If TRUE add an extension for flank transmission.}

\item{flank_impact}{One parameter for the flank extension.}

\item{flank_power}{Another parameter for the flank extension.}

\item{flank_start}{At which frequency should the flank transmission apply?}

\item{freq}{Vector of frequencies.}

\item{sts}{If TRUE the model is calculated strict according to standard ISO 12354-1. See strict_to_standard() for more information.}

\item{max_rl}{Maximal radiation level value}

\item{max_rl_type}{Maximum radiation level type. See ?smooth_radiation_level.}

\item{offset}{R+offset}

\item{...}{Further arguments mainly from the function for the radiation level.}
}
\value{
Soundproofing measure R
}
\description{
Calculate the soundproofing measure R
}
\examples{
soundproofing_measure()
}
