% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_data_list_inkl_cov.R
\name{spm_data_list_inkl_cov}
\alias{spm_data_list_inkl_cov}
\title{spm_data_list_inkl_cov: Title}
\usage{
spm_data_list_inkl_cov(data, methods = c("cov", "lwls", "var", "mean"))
}
\arguments{
\item{data}{data.frame with sound reduction index values}

\item{methods}{Details will follow, but you can specify here which estimators are to be calculated.}
}
\description{
A description of spm_data_list_inkl_cov
}
\details{
A details of spm_data_list_inkl_cov
}
