% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strict_to_standard.R
\name{strict_to_standard}
\alias{strict_to_standard}
\title{What should be taken into account if the calculations are to be carried out strictly in accordance with the standard?}
\usage{
strict_to_standard()
}
\description{
What should be taken into account if the calculations are to be carried out strictly in accordance with the standard?
}
