% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_norm_para.R
\name{trunc_norm_para}
\alias{trunc_norm_para}
\title{Find µ and \eqn{\sigma} for a truncated gaussian given mean and standard deviation}
\usage{
trunc_norm_para(low, upp, mean, sd)
}
\arguments{
\item{low}{lower bound}

\item{upp}{upper bound}

\item{mean}{mean value of the truncated gaussian}

\item{sd}{standarad deviation of the truncated gaussian}
}
\description{
Find µ and \eqn{\sigma} for a truncated gaussian given mean and standard deviation
}
\details{
For the optimization see ?nleqslv::nleqslv
}
