% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvARcov2coeff.R
\name{tvARcov2coeff}
\alias{tvARcov2coeff}
\title{transforms a tvAR covariance matrix into a matrix of ar coefficients}
\usage{
tvARcov2coeff(x)
}
\arguments{
\item{x}{tvAR covariance matrix}
}
\description{
transforms a tvAR covariance matrix into a matrix of ar coefficients
}
