% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_to_side_length.R
\name{volume_to_side_length}
\alias{volume_to_side_length}
\title{Converts the volume of a wall into a side length}
\usage{
volume_to_side_length(
  volume,
  height = 2.5,
  side_ratio = 0.8,
  output = c("longer", "shorter", "both", "perimeter")
)
}
\arguments{
\item{volume}{m³}

\item{height}{m}

\item{output}{choose return: longer, shorter, both, perimeter}

\item{side_ration}{short_side/long_side}
}
\description{
Converts the volume of a wall into a side length
}
