% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiteningTVAR.R
\name{whiteningTVAR}
\alias{whiteningTVAR}
\title{Performs a whitening under the TVAR assumption}
\usage{
whiteningTVAR(x, maxlag, returnSS = T)
}
\arguments{
\item{x}{a \eqn{I \times J} matrix}

\item{maxlag}{maximum number of lags}

\item{returnSS}{return squared sum for each row}
}
\description{
Performs a whitening under the TVAR assumption
}
